requirements.txt
# Backend dependencies for Databricks Apps
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
databricks-sdk==0.18.0
python-multipart==0.0.6
pyyaml==6.0.1

# Enterprise Platform Dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.8.2
